blacklist: Use word boundaries in blacklist
Compile all slugified words as (escaped) regular expressions with word
boundaries and use pattern matching to scan the message rather than
using the more naive `substr in str` syntax.

This helps ensure that the words which are banned are actually being
used in words, i.e. the following blacklist

 - niga

Does not match the following message

  "I've had enough of these shenanigans"